fix: resolve all clippy warnings, migrate png→zenpng
- Add zenpng workspace dep, replace all png::Decoder usage in tests/benches/bench-utils
  with zenjpeg_bench_utils::load_png()/load_png_bytes() backed by zenpng
- Fix butteraugli deprecation: compute_butteraugli() → butteraugli::butteraugli() with ImgRef<RGB8>
- Replace criterion::black_box with std::hint::black_box in all benchmarks
- Collapse nested ifs with Rust 2024 let chains across lib/tests/examples
- Fix #[arcane] unused_assignments with `let _ = pos;` in idct_int.rs
- Remove dead code (scans_decoded in parser), add #[allow(dead_code)] where needed
- Gate rst_scan module behind #[cfg(feature = "parallel")]
- Fix Rust 2024 unsafe block requirements in jpegli-internals-sys
- Fix &PathBuf → &Path, &mut Vec → &mut [T], .map(|s| s.clone()) → .cloned()
- 110 files changed, net -130 lines

+++ zenjpeg/tests/mcu_border_roundtrip.rs
@@ -149,7 +149,7 @@ fn mozjpeg_to_zenjpeg(width: u32, height: u32, ss: mozjpeg_rs::Subsampling) {
     let jpeg = mozjpeg_rs::Encoder::new(mozjpeg_rs::Preset::BaselineBalanced)
         .quality(90)
         .subsampling(ss)
-        .encode_rgb(&img.pixels, width as u32, height as u32)
+        .encode_rgb(&img.pixels, width, height)
         .unwrap_or_else(|e| panic!("mozjpeg encode {}x{}: {e}", width, height));
 
     let decoder = Decoder::new();
